+++ old-code-temp/src copy/components/EarmarkFooter/index.tsx
@@ -1,9 +1,0 @@
-/* eslint-disable */
-import Link from 'next/link'
-const EarmarkFooter = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default EarmarkFooter;