+++ old-code-temp/src copy/components/Nav/TabNav/index.tsx
@@ -1,4 +1,0 @@
-/* eslint-disable */
-const TabNav = () => {};
-
-export default TabNav;