+++ old-code-temp/src copy/lib/parseCookies/index.ts
@@ -1,6 +1,0 @@
-/* eslint-disable */
-import cookie from "cookie"
-
-export function parseCookies(req) {
-  return cookie.parse(req ? req.headers.cookie || "" : document.cookie)
-}